IceFloeTracker

Documentation for IceFloeTracker.jl package.

IceFloeTracker

Functions

Index

Functions

IceFloeTracker._bin9todec — Method

_bin9todec(v)

Get decimal representation of a bit vector v with the leading bit at its leftmost posistion.

Example

julia> _bin9todec([0 0 0 0 0 0 0 0 0])    
0

julia> _bin9todec([1 1 1 1 1 1 1 1 1])    
511

IceFloeTracker._operator_lut — Method

_operator_lut(I, img, nhood, lut1, lut2)

Look up the neighborhood nhood in lookup tables lut1 and lut2.

Handles cases when the center of nhood is on the edge of img using data in I.

IceFloeTracker.add_padding — Method

add_padding(img, style)

Extrapolate the image img according to the style specifications type. Returns the extrapolated image.

Arguments

img: Image to be padded.

style: A supported type (such as Pad or Fill) representing the extrapolation style. See the relevant documentation for details.

See also remove_padding

IceFloeTracker.addfloemasks! — Method

addfloearrays(props::DataFrame, floeimg::BitMatrix)

Add a column to props called floearray containing the cropped floe masks from floeimg.

IceFloeTracker.bwareamaxfilt — Function

bwareamaxfilt(bwimg::AbstractArray{Bool}, conn)

Filter the smaller (by area) connected components in bwimg keeping the (assumed unique) largest.

Uses 8-pixel connectivity by default (conn=8). Use conn=4 for 4-pixel connectivity.

IceFloeTracker.bwareamaxfilt! — Function

bwareamaxfilt!(bwimg::AbstractArray)

In-place version of bwareamaxfilt.

See also bwareamaxfilt

IceFloeTracker.bwdist — Method

bwdist(bwimg)

Distance transform for binary image bwdist.

IceFloeTracker.check_fname — Function

check_fname(fname)

Checks fname does not exist in current directory; throws an assertion if this condition is false.

Arguments

fname: String object or Symbol to a reference to a String representing a path.

IceFloeTracker.cropfloe — Method

cropfloe(floesimg, props, i)

Crops the floe delimited by the bounding box data in props at index i from the floe image floesimg.

IceFloeTracker.filt_except_label! — Method

filt_except_label!(labeled_arr::Array{Int64, 2}, label::Int64)

In-place version of filt_except_label.

See also filt_except_label

IceFloeTracker.filt_except_label — Method

filt_except_label(labeled_arr::Array{Int64, 2}, label::Int64)

Make 0 all values in labeled_arr that are not equal to label.

See also filt_except_label!

IceFloeTracker.fixzeroindexing! — Method

fixzeroindexing!(props::DataFrame, props_to_fix::Vector{T}) where T<:Union{Symbol,String}

Fix the zero-indexing of the props_to_fix columns in props by adding 1 to each element.

IceFloeTracker.fname_ext_splice — Method

fname_ext_splice(fname, ext)

Join "fname" and "ext" with '.'.

IceFloeTracker.fname_ext_split — Method

fname_ext_split(fname)

Split "fname.ext" into "fname" and "ext".

IceFloeTracker.get_areas — Method

get_areas(labeled_arr::Array{T, 2})::Dict{T, Int} where T

Get the "areas" (count of pixels of a given label) of the connected components in labeled_arr.

Return a dictionary with the frequency distribution: label => countoflabel.

IceFloeTracker.get_max_label — Method

get_max_label(d::Dict{Int, Int})

Get the label k in dictionary d for which d[k] is maximal.

IceFloeTracker.getbboxcolumns — Method

getbboxcolumns(props::DataFrame)

Return the column names of the bounding box columns in props.

IceFloeTracker.getcentroidcolumns — Method

getcentroidcolumns(props::DataFrame)

Returns the column names of the centroid columns in props.

IceFloeTracker.getfloemasks — Method

getfloemasks(props::DataFrame, floeimg::BitMatrix)

Return a vector of cropped floe masks from floeimg using the bounding box data in props.

IceFloeTracker.imextendedmin — Method

imextendedmin(img)

Mimics MATLAB's imextendedmin function that computes the extended-minima transform, which is the regional minima of the H-minima transform. Regional minima are connected components of pixels with a constant intensity value. This function returns a transformed bitmatrix.

Arguments

img: image object

h: suppress minima below this depth threshold

conn: neighborhood connectivity; in 2D 1 = 4-neighborhood and 2 = 8-neighborhood

IceFloeTracker.loadimg — Method

loadimg(; dir::String, fname::String)

Load an image from dir with filename fname into a matrix of Float64 values. Returns the loaded image.

IceFloeTracker.make_filename — Method

make_filename()

Makes default filename with timestamp.

IceFloeTracker.padnhood — Method

padnhood(img, I, nhood)

Pad the matrix img[nhood] with zeros according to the position of I within the edgesimg.

Returns img[nhood] if I is not an edge index.

IceFloeTracker.regionprops — Function

regionprops(label_img, ; properties, connectivity)

A wrapper of the regionprops function from the skimage python library.

See its full documentation at https://scikit-image.org/docs/stable/api/skimage.measure.html#skimage.measure.regionprops.

Arguments

label_img: Image with the labeled objects of interest

intensity_img: (Optional) Used for generating extra_properties, integer/float array from which (presumably) label_img was generated

extra_properties: (Optional) not yet implemented. It will be set to nothing

See also regionprops_table

Examples

julia> Random.seed!(123);

julia> bw_img = rand([0, 1], 5, 10)
5×10 Matrix{Int64}:
 1  0  1  0  0  0  0  0  0  1
 1  0  1  1  1  0  0  0  1  1
 1  1  0  1  1  0  1  0  0  1
 0  1  0  1  0  0  0  0  1  0
 1  0  0  0  0  1  0  1  0  1

 julia> label_img = Images.label_components(bw_img, trues(3,3))
 5×10 Matrix{Int64}:
  1  0  1  0  0  0  0  0  0  4
  1  0  1  1  1  0  0  0  4  4
  1  1  0  1  1  0  3  0  0  4
  0  1  0  1  0  0  0  0  4  0
  1  0  0  0  0  2  0  4  0  4

 julia> regions = regionprops(label_img, bw_img);

 julia> for region in regions
           println(region.area,"	", region.perimeter)
        end
13      11.621320343559642
1       0.0
1       0.0
7       4.621320343559642

IceFloeTracker.regionprops_table — Function

regionprops_table(label_img, intensity_img; properties, connectivity, extra_properties)

A wrapper of the regionprops_table function from the skimage python library.

See its full documentation at https://scikit-image.org/docs/stable/api/skimage.measure.html#regionprops-table.

Arguments

label_img: Image with the labeled objects of interest

intensity_img: (Optional) Used for generating extra_properties, integer/float array from which (presumably) label_img was generated

properties: List (Vector or Tuple) of properties to be generated for each connected component in label_img

extra_properties: (Optional) not yet implemented. It will be set to nothing

Notes

Zero indexing has been corrected for the bbox and centroid properties

bbox data (max_col and max_row) are inclusive

centroid data are rounded to the nearest integer

See also regionprops

Examples

julia> using IceFloeTracker, Random

julia> Random.seed!(123);

julia> bw_img = rand([0, 1], 5, 10)
5×10 Matrix{Int64}:
 1  0  1  0  0  0  0  0  0  1
 1  0  1  1  1  0  0  0  1  1
 1  1  0  1  1  0  1  0  0  1
 0  1  0  1  0  0  0  0  1  0
 1  0  0  0  0  1  0  1  0  1

julia> label_img = IceFloeTracker.label_components(bw_img, trues(3,3))
5×10 Matrix{Int64}:
 1  0  1  0  0  0  0  0  0  4
 1  0  1  1  1  0  0  0  4  4
 1  1  0  1  1  0  3  0  0  4
 0  1  0  1  0  0  0  0  4  0
 1  0  0  0  0  2  0  4  0  4

julia> properties = ["area", "perimeter"]
2-element Vector{String}:
 "area"
 "perimeter"

 julia> IceFloeTracker.regionprops_table(label_img, bw_img, properties = properties)
 4×2 DataFrame
  Row │ area   perimeter 
      │ Int32  Float64   
 ─────┼──────────────────
    1 │    13   11.6213
    2 │     1    0.0
    3 │     1    0.0
    4 │     7    4.62132

IceFloeTracker.remove_padding — Method

remove_padding(paddedimg, border_spec)

Removes padding from the boundary of padded image paddedimg according to the border specification border_spec type. Returns the cropped image.

Arguments

paddedimg: Pre-padded image.

border_spec: Type representing the style of padding (such as Pad or Fill) with which paddedimg is assumend to be pre-padded. Example: Pad((1,2), (3,4)) specifies 1 row on the top, 2 columns on the left, 3 rows on the bottom, and 4 columns on the right boundary.

See also add_padding

IceFloeTracker.renamecols! — Method

renamecols!(props::DataFrame, oldnames::Vector{T}, newnames::Vector{T}) where T<:Union{Symbol,String}

Rename the oldnames columns in props to newnames.

IceFloeTracker.roundtoint! — Method

roundtoint!(props::DataFrame, colnames::Vector{T}) where T<:Union{Symbol,String}

Round the colnames columns in props to Int.

IceFloeTracker.timestamp — Method

timestamp(fname)

Attach timestamp to fname.

IceFloeTracker.@persist — Macro

@persist img fname
@persist(img,fname)
@persist img
@persist(img)
@persist img fname ts
@persist(img, fname, ts)

Given a reference to an image object img, the macro persists (saves to a file) img to the current working directory using fname as filename. Returns img.

Arguments

img: Symbol expression representing an image object loaded in memory.

fname: Optional filename for the persisted image.

ts: Optional boolean to attach timestamp to fname.

Index

IceFloeTracker._bin9todec

IceFloeTracker._operator_lut

IceFloeTracker.add_padding

IceFloeTracker.addfloemasks!

IceFloeTracker.bwareamaxfilt

IceFloeTracker.bwareamaxfilt!

IceFloeTracker.bwdist

IceFloeTracker.check_fname

IceFloeTracker.cropfloe

IceFloeTracker.filt_except_label

IceFloeTracker.filt_except_label!

IceFloeTracker.fixzeroindexing!

IceFloeTracker.fname_ext_splice

IceFloeTracker.fname_ext_split

IceFloeTracker.get_areas

IceFloeTracker.get_max_label

IceFloeTracker.getbboxcolumns

IceFloeTracker.getcentroidcolumns

IceFloeTracker.getfloemasks

IceFloeTracker.imextendedmin

IceFloeTracker.loadimg

IceFloeTracker.make_filename

IceFloeTracker.padnhood

IceFloeTracker.regionprops

IceFloeTracker.regionprops_table

IceFloeTracker.remove_padding

IceFloeTracker.renamecols!

IceFloeTracker.roundtoint!

IceFloeTracker.timestamp

IceFloeTracker.@persist