IceFloeTracker

Documentation for IceFloeTracker.jl package.

IceFloeTracker

Functions

Index

Functions

IceFloeTracker.add_padding — Method

add_padding(img, style)

Extrapolate the image img according to the style specifications type. Returns the extrapolated image.

Arguments

img: Image to be padded.

style: A supported type (such as Pad or Fill) representing the extrapolation style. See the relevant documentation for details.

See also remove_padding

IceFloeTracker.apply_landmask! — Method

apply_landmask(input_image, landmask_binary)

Zero out pixels in land and soft ice regions on truecolor image, return RGB image with zero for all three channels on land/soft ice.

Arguments

input_image: truecolor RGB image

landmask_binary: binary landmask with 1=land, 0=water/ice

IceFloeTracker.check_fname — Function

check_fname(fname)

Checks fname does not exist in current directory; throws an assertion if this condition is false.

Arguments

fname: String object or Symbol to a reference to a String representing a path.

IceFloeTracker.create_landmask — Method

create_landmask(landmask_image, num_pixels_dilate, fill_value)

Convert a 3-channel RGB land mask image to a 1-channel binary matrix, including a buffer to extend the land over any soft ice regions; land = 0, water/ice = 1.

Arguments

landmask_image: RGB or boolean image with land pixels represented as true or 1

struct_elem: structuring element for dilation

fill_value: number of pixels used to fill holes in land mask

IceFloeTracker.landmask — Method

landmask(;metadata, input, output)

Given an input directory with a landmask file and truecolor images, create a land/soft ice mask, and apply to the truecolor images. The resulting land-masked images are saved to the snakemake output directory.

Arguments

metadata: JSON file with metadata

input: path to image dir containing truecolor and landmask images

output: path to output dir where land-masked truecolor images are saved

IceFloeTracker.make_filename — Method

make_filename()

Makes default filename with timestamp.

IceFloeTracker.remove_padding — Method

remove_padding(paddedimg, border_spec)

Removes padding from the boundary of padded image paddedimg according to the border specification border_spec type. Returns the cropped image.

Arguments

paddedimg: Pre-padded image.

border_spec: Type representing the style of padding (such as Pad or Fill) with which paddedimg is assumend to be pre-padded. Example: Pad((1,2), (3,4)) specifies 1 row on the top, 2 columns on the left, 3 rows on the bottom, and 4 columns on the right boundary.

See also add_padding

IceFloeTracker.@persist — Macro

@persist img fname or
@persist(img,fname) or
@persist img or
@persist(img)

Given a reference to an image object img, the macro persists (saves to a file) img to the current working directory using fname as filename. Returns img.

Arguments

img: Symbol expression representing an image object loaded in memory.

fname: Optional filename for the persisted image.

Index

IceFloeTracker.add_padding

IceFloeTracker.apply_landmask!

IceFloeTracker.check_fname

IceFloeTracker.create_landmask

IceFloeTracker.landmask

IceFloeTracker.make_filename

IceFloeTracker.remove_padding

IceFloeTracker.@persist