IceFloeTracker

Documentation for IceFloeTracker.jl package.

IceFloeTracker

Functions

Index

Functions

IceFloeTracker.add_padding — Method

add_padding(img, style)

Extrapolate the image img according to the style specifications type. Returns the extrapolated image.

Arguments

img: Image to be padded.

style: A supported type (such as Pad or Fill) representing the extrapolation style. See the relevant documentation for details.

See also remove_padding

IceFloeTracker.apply_cloudmask — Method

apply_cloudmask(reflectance_image, cloudmask)

Zero out pixels containing clouds where clouds and ice are not discernable. Arguments should be of the same size.

Arguments

reflectance_image: corrected reflectance false color image - bands [7,2,1]

cloudmask: binary cloudmask with clouds = 0, else = 1

IceFloeTracker.apply_landmask! — Method

apply_landmask(input_image, landmask_binary)

Zero out pixels in land and soft ice regions on truecolor image, return RGB image with zero for all three channels on land/soft ice.

Arguments

input_image: truecolor RGB image

landmask_binary: binary landmask with 1=land, 0=water/ice

IceFloeTracker.bwperim — Function

bwperim(bwimg, conn::Int=4)

Locate the pixels at the boundary of objects in an binary image bwimg using connectivity conn.

Arguments

bwimg: Binary (black/white – 1/0) image

conn: connectivity. Either 4 or 8 for 4-pixel and 8-pixel connectivity, respectively.

Examples

julia> A = zeros(Int, 13, 16); A[2:6, 2:6] .= 1; A[4:8, 7:10] .= 1; A[10:12,13:15] .= 1; A[10:12,3:6] .= 1;

julia> A
13×16 Matrix{Int64}:
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0

 julia> bwperim(A)
13×16 Matrix{Bool}:
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  0  0  0  1  0  0  0  0  0  0  0  0  0  0
 0  1  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  1  0  0  0  0  0  0  0  1  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  0  0  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  0  0  1  0  0  0  0  0  0  1  0  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0

IceFloeTracker.check_fname — Function

check_fname(fname)

Checks fname does not exist in current directory; throws an assertion if this condition is false.

Arguments

fname: String object or Symbol to a reference to a String representing a path.

IceFloeTracker.create_cloudmask — Method

create_cloudmask(reflectance_image)

Convert a 3-channel false color reflectance image to a 1-channel binary matrix; clouds = 0, else = 1. Default thresholds are defined in the published Ice Floe Tracker article: Remote Sensing of the Environment 234 (2019) 111406.

Arguments

ref_image: corrected reflectance false color image - bands [7,2,1]

prelim_threshold: threshold value used to identify clouds in band 7, N0f8(RGB intensity/255)

band_7_threshold: threshold value used to identify cloud-ice in band 7, N0f8(RGB intensity/255)

band_2_threshold: threshold value used to identify cloud-ice in band 2, N0f8(RGB intensity/255)

ratio_lower: threshold value used to set lower ratio of cloud-ice in bands 7 and 2

ratio_upper: threshold value used to set upper ratio of cloud-ice in bands 7 and 2

IceFloeTracker.create_landmask — Method

create_landmask(landmask_image, num_pixels_dilate, fill_value)

Convert a 3-channel RGB land mask image to a 1-channel binary matrix, including a buffer to extend the land over any soft ice regions; land = 0, water/ice = 1.

Arguments

landmask_image: RGB or boolean image with land pixels represented as true or 1

struct_elem: structuring element for dilation

fill_value: number of pixels used to fill holes in land mask

IceFloeTracker.discriminate_ice_water — Method

discriminate_ice_water(reflectance_image, reflectance_image_band7, normalized_image, landmask_bitmatrix, clouds_channel; floes_threshold, mask_clouds_lower, mask_clouds_upper, kurt_thresh_lower, kurt_thresh_upper, skew_thresh, st_dev_thresh_lower, st_dev_thresh_upper, clouds2_threshold, differ_threshold, nbins)

Generates an image with ice floes apparent after filtering and combining previously processed versions of reflectance and truecolor images from the same region of interest. Returns an image ready for segmentation to isolate floes.

Arguments

reflectance_image: input image in false color reflectance

reflectance_image_band7: first channel (band 7) of reflectance image, output from cloudmask.jl

normalized_image: a normalized, landmasked truecolor image, output from normalization.jl

landmask_bitmatrix: bitmatrix landmask for region of interest

clouds_channel: first channel of cloudmasked reflectance image, output from cloudmask.jl

floes_threshold: heuristic applied to original reflectance image

mask_clouds_lower: lower heuristic applied to mask out clouds

mask_clouds_upper: upper heuristic applied to mask out clouds

kurt_thresh_lower: lower heuristic used to set pixel value threshold based on kurtosis in histogram

kurt_thresh_upper: upper heuristic used to set pixel value threshold based on kurtosis in histogram

skew_thresh: heuristic used to set pixel value threshold based on skewness in histogram

st_dev_thresh_lower: lower heuristic used to set pixel value threshold based on standard deviation in histogram

st_dev_thresh_upper: upper heuristic used to set pixel value threshold based on standard deviation in histogram

clouds2_threshold: heuristic used to set pixel value threshold based on ratio of clouds

differ_threshold: heuristic used to calculate proportional intensity in histogram

nbins: number of bins during histogram build

verify then add all the defaulted params

IceFloeTracker.landmask — Method

landmask(;metadata, input, output)

Given an input directory with a landmask file and truecolor images, create a land/soft ice mask, and apply to the truecolor images. The resulting land-masked images are saved to the snakemake output directory.

Arguments

metadata: JSON file with metadata

input: path to image dir containing truecolor and landmask images

output: path to output dir where land-masked truecolor images are saved

IceFloeTracker.make_filename — Method

make_filename()

Makes default filename with timestamp.

IceFloeTracker.normalize_image — Method

normalize_image(truecolor_image, landmask, struct_elem; lambda, kappa, niters, nbins, rblocks, cblocks, clip, smoothing_param, intensity)

Adjusts truecolor land-masked image to highlight ice floe features. This function performs diffusion, adaptive histogram equalization, and sharpening, and returns a greyscale normalized image.

Arguments

truecolor_image: input image in truecolor

landmask: bitmatrix landmask for region of interest

struct_elem: structuring element for dilation

lambda: speed of diffusion (0–0.25)

kappa: conduction coefficient for diffusion (25–100)

niters: number of iterations of diffusion

nbins: number of bins during histogram equalization

rblocks: number of row blocks to divide input image during equalization

cblocks: number of column blocks to divide input image during equalization

clip: threshold for clipping histogram bins (0–1); values closer to one minimize contrast enhancement, values closer to zero maximize contrast enhancement

smoothing_param: pixel radius for gaussian blurring (1–10)

intensity: amount of sharpening to perform

IceFloeTracker.remove_padding — Method

remove_padding(paddedimg, border_spec)

Removes padding from the boundary of padded image paddedimg according to the border specification border_spec type. Returns the cropped image.

Arguments

paddedimg: Pre-padded image.

border_spec: Type representing the style of padding (such as Pad or Fill) with which paddedimg is assumend to be pre-padded. Example: Pad((1,2), (3,4)) specifies 1 row on the top, 2 columns on the left, 3 rows on the bottom, and 4 columns on the right boundary.

See also add_padding

IceFloeTracker.@persist — Macro

@persist img fname or
@persist(img,fname) or
@persist img or
@persist(img)

Given a reference to an image object img, the macro persists (saves to a file) img to the current working directory using fname as filename. Returns img.

Arguments

img: Symbol expression representing an image object loaded in memory.

fname: Optional filename for the persisted image.

Index

IceFloeTracker.add_padding

IceFloeTracker.apply_cloudmask

IceFloeTracker.apply_landmask!

IceFloeTracker.bwperim

IceFloeTracker.check_fname

IceFloeTracker.create_cloudmask

IceFloeTracker.create_landmask

IceFloeTracker.discriminate_ice_water

IceFloeTracker.landmask

IceFloeTracker.make_filename

IceFloeTracker.normalize_image

IceFloeTracker.remove_padding

IceFloeTracker.@persist