IceFloeTracker

Documentation for IceFloeTracker.jl package.

IceFloeTracker

Functions

Index

Functions

IceFloeTracker.add_padding — Method

add_padding(img, style)

Extrapolate the image img according to the style specifications type. Returns the extrapolated image.

Arguments

img: Image to be padded.

style: A supported type (such as Pad or Fill) representing the extrapolation style. See the relevant documentation for details.

See also remove_padding

IceFloeTracker.apply_cloudmask — Method

apply_cloudmask(reflectance_image, cloudmask)

Zero out pixels containing clouds where clouds and ice are not discernable. Arguments should be of the same size.

Arguments

reference_image: corrected reflectance false color image - bands [7,2,1] or grayscale

cloudmask: binary cloudmask with clouds = 0, else = 1

IceFloeTracker.apply_landmask — Method

apply_landmask(input_image, landmask_binary)

Zero out pixels in land and soft ice regions on truecolor image, return RGB image with zero for all three channels on land/soft ice.

Arguments

input_image: truecolor RGB image

landmask_binary: binary landmask with 1=land, 0=water/ice

IceFloeTracker.atan2 — Method

atan2(y,x)

Wrapper of Base.atan that returns the angle of vector (x,y) in the range [0, 2π).

IceFloeTracker.bwareamaxfilt! — Method

bwareamaxfilt!(bwimg::AbstractArray)

In-place version of bwareamaxfilt.

See also bwareamaxfilt

IceFloeTracker.bwareamaxfilt — Method

bwareamaxfilt(bwimg::AbstractArray{Bool})

Filter the smaller (by area) connected components in bwimg keeping the (assumed unique) largest.

IceFloeTracker.bwperim — Method

bwperim(bwimg)

Locate the pixels at the boundary of objects in an binary image bwimg using 8-pixel connectivity.

Arguments

bwimg: Binary (black/white – 1/0) image

Examples

julia> A = zeros(Int, 13, 16); A[2:6, 2:6] .= 1; A[4:8, 7:10] .= 1; A[10:12,13:15] .= 1; A[10:12,3:6] .= 1;

julia> A
13×16 Matrix{Int64}:
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0

 julia> bwperim(A)
13×16 Matrix{Bool}:
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  0  0  0  1  0  0  0  0  0  0  0  0  0  0
 0  1  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  1  0  0  0  0  0  0  0  1  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  0  0  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  0  0  1  0  0  0  0  0  0  1  0  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0

IceFloeTracker.bwtraceboundary — Method

bwtraceboundary(image::Union{Matrix{Int64},Matrix{Float64},BitMatrix};
                P0::Union{Tuple{Int,Int},CartesianIndex{2},Nothing}=nothing,
                closed::Bool=true)

Trace the boundary of objects in image

Background pixels are represented as zero. The algorithm traces the boundary counterclockwise and an initial point P0 can be specified. If more than one boundary is detected and an initial point is provided, the boundary that contains this point is returned as a vector of CartesianIndex types. Otherwise an array of vectors is returned with all the detected boundaries in image.

Arguments

image: image, preferably binary with one single object, whose objects' boundaries are to be traced.

P0: initial point of a target boundary.

closed: if true (default) makes the inital point of a boundary equal to the last point.

Example

julia> A = zeros(Int, 13, 16); A[2:6, 2:6] .= 1; A[4:8, 7:10] .= 1; A[10:12,13:15] .= 1; A[10:12,3:6] .= 1;

julia> A
13×16 Matrix{Int64}:
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0

julia> boundary = IceFloeTracker.bwtraceboundary(A);

julia> boundary[3]
9-element Vector{CartesianIndex}:
 CartesianIndex(10, 13)
 CartesianIndex(11, 13)
 CartesianIndex(12, 13)
 CartesianIndex(12, 14)
 CartesianIndex(12, 15)
 CartesianIndex(11, 15)
 CartesianIndex(10, 15)
 CartesianIndex(10, 14)
 CartesianIndex(10, 13)

IceFloeTracker.check_fname — Function

check_fname(fname)

Checks fname does not exist in current directory; throws an assertion if this condition is false.

Arguments

fname: String object or Symbol to a reference to a String representing a path.

IceFloeTracker.clockwise — Method

Make a clockwise turn from the dir direction

IceFloeTracker.counterclockwise — Method

Make a counterclockwise turn from the dir direction

IceFloeTracker.create_cloudmask — Method

create_cloudmask(reflectance_image; prelim_threshold, band_7_threshold, band_2_threshold, ratio_lower, ratio_upper)

Convert a 3-channel false color reflectance image to a 1-channel binary matrix; clouds = 0, else = 1. Default thresholds are defined in the published Ice Floe Tracker article: Remote Sensing of the Environment 234 (2019) 111406.

Arguments

reflectance_image: corrected reflectance false color image - bands [7,2,1]

prelim_threshold: threshold value used to identify clouds in band 7, N0f8(RGB intensity/255)

band_7_threshold: threshold value used to identify cloud-ice in band 7, N0f8(RGB intensity/255)

band_2_threshold: threshold value used to identify cloud-ice in band 2, N0f8(RGB intensity/255)

ratio_lower: threshold value used to set lower ratio of cloud-ice in bands 7 and 2

ratio_upper: threshold value used to set upper ratio of cloud-ice in bands 7 and 2

IceFloeTracker.create_landmask — Method

create_landmask(landmask_image, num_pixels_dilate, fill_value)

Convert a 3-channel RGB land mask image to a 1-channel binary matrix, including a buffer to extend the land over any soft ice regions; land = 0, water/ice = 1.

Arguments

landmask_image: land mask image

struct_elem: structuring element for dilation

fill_value: number of pixels used to fill holes in land mask

IceFloeTracker.crosscorr — Method

r, lags = crosscorr(u::Vector{T},
                    v::Vector{T};
                    normalize::Bool=false,
                    padmode::Symbol=:longest)

Wrapper of DSP.xcorr with normalization (see https://docs.juliadsp.org/stable/convolutions/#DSP.xcorr)

Returns the pair (r, lags) with the cross correlation scores r and corresponding lags according to padmode.

Arguments

u,v: real vectors which could have unequal length.

normalize: return normalized correlation scores (false by default).

padmode: either :longest (default) or :none to control padding of shorter vector with zeros.

Examples

The example below builds two vectors, one a shifted version of the other, and computes various cross correlation scores.

julia> n = 1:5;

julia> x = 0.48.^n;

julia> y = circshift(x,3);

julia> r, lags = crosscorr(x,y,normalize=true);

julia> [r lags]
9×2 Matrix{Float64}:
0.369648    -4.0
0.947531    -3.0
0.495695    -2.0
0.3231      -1.0
0.332519     0.0
0.15019      1.0
0.052469     2.0
0.0241435    3.0
0.00941878   4.0

julia> r, lags = crosscorr(x,y,normalize=true,padmode=:none);

julia> [r lags]
9×2 Matrix{Float64}:
0.369648    1.0
0.947531    2.0
0.495695    3.0
0.3231      4.0
0.332519    5.0
0.15019     6.0
0.052469    7.0
0.0241435   8.0
0.00941878  9.0

This final example builds two vectors of different length and computes some cross correlation scores.

julia> n = 1:5; m = 1:3;

julia> x = 0.48.^n; y = 0.48.^m;

julia> r, lags = crosscorr(x,y,normalize=true);

julia> [r lags]
9×2 Matrix{Float64}:
0.0          -4.0
4.14728e-17  -3.0
0.178468     -2.0
0.457473     -1.0
0.994189      0.0
0.477211      1.0
0.229061      2.0
0.105402      3.0
0.0411191     4.0

julia> r, lags = crosscorr(x,y,normalize=true,padmode=:none);

julia> [r lags]
7×2 Matrix{Float64}:
0.178468   1.0
0.457473   2.0
0.994189   3.0
0.477211   4.0
0.229061   5.0
0.105402   6.0
0.0411191  7.0

IceFloeTracker.detect_move! — Method

Workhorse function: Get all pixel coords for detected border.

IceFloeTracker.discriminate_ice_water — Method

discriminate_ice_water(reflectance_image, reflectance_image_band7, normalized_image, landmask_bitmatrix, clouds_channel; floes_threshold, mask_clouds_lower, mask_clouds_upper, kurt_thresh_lower, kurt_thresh_upper, skew_thresh, st_dev_thresh_lower, st_dev_thresh_upper, clouds2_threshold, differ_threshold, nbins)

Generates an image with ice floes apparent after filtering and combining previously processed versions of reflectance and truecolor images from the same region of interest. Returns an image ready for segmentation to isolate floes.

Arguments

reflectance_image: input image in false color reflectance

reflectance_image_band7: first channel (band 7) of reflectance image, output from cloudmask.jl

normalized_image: a normalized, landmasked truecolor image, output from normalization.jl

landmask_bitmatrix: bitmatrix landmask for region of interest

clouds_channel: first channel of cloudmasked reflectance image, output from cloudmask.jl

floes_threshold: heuristic applied to original reflectance image

mask_clouds_lower: lower heuristic applied to mask out clouds

mask_clouds_upper: upper heuristic applied to mask out clouds

kurt_thresh_lower: lower heuristic used to set pixel value threshold based on kurtosis in histogram

kurt_thresh_upper: upper heuristic used to set pixel value threshold based on kurtosis in histogram

skew_thresh: heuristic used to set pixel value threshold based on skewness in histogram

st_dev_thresh_lower: lower heuristic used to set pixel value threshold based on standard deviation in histogram

st_dev_thresh_upper: upper heuristic used to set pixel value threshold based on standard deviation in histogram

clouds2_threshold: heuristic used to set pixel value threshold based on ratio of clouds

differ_threshold: heuristic used to calculate proportional intensity in histogram

nbins: number of bins during histogram build

IceFloeTracker.filt_except_label! — Method

filt_except_label!(labeled_arr::Array{Int64, 2}, label::Int64)

In-place version of filt_except_label.

See also filt_except_label

IceFloeTracker.filt_except_label — Method

filt_except_label(labeled_arr::Array{Int64, 2}, label::Int64)

Make 0 all values in labeled_arr that are not equal to label.

See also filt_except_label!

IceFloeTracker.from_to — Method

Get index in Moore neigborhood representing the direction from the from pixel coords to the to pixel coords (see definition of dir_delta below).

Clockwise Moore neighborhood.

dir_delta = [CartesianIndex(-1, 0), CartesianIndex(-1, 1), CartesianIndex(0, 1), CartesianIndex(1, 1), CartesianIndex(1, 0), CartesianIndex(1, -1), CartesianIndex(0, -1), CartesianIndex(-1,-1)]

IceFloeTracker.get_areas — Method

get_areas(labeled_arr::Array{T, 2})::Dict{T, Int} where T

Get the "areas" (count of pixels of a given label) of the connected components in labeled_arr.

Return a dictionary with the frequency distribution: label => countoflabel.

IceFloeTracker.get_max_label — Method

get_max_label(d::Dict{Int, Int})

Get the label k in dictionary d for which d[k] is maximal.

IceFloeTracker.grad — Method

dx, dy = grad(A::Matrix{<:Number})

Make gradient vector field for the set of points with coordinates in the rows of the matrix A with x-coordinates down column 1 and y-coordinates down column 2. Return a tuple with dx and dy in that order.

IceFloeTracker.grad — Method

dx, dy = grad(x::Vector{<:Number}, y::Vector{<:Number})

Make gradient vector field for the set of points with coordinates in vectors x and y. Return a tuple with dx and dy in that order.

IceFloeTracker.isincountourlist — Method

Check P is in countour list if so return the index of the contour that contains P, otherwise return false.

IceFloeTracker.landmask — Method

landmask(;metadata, input, output)

Given an input directory with a landmask file and truecolor images, create a land/soft ice mask, and apply to the truecolor images. The resulting land-masked images are saved to the snakemake output directory.

Arguments

metadata: JSON file with metadata

input: path to image dir containing truecolor and landmask images

output: path to output dir where land-masked truecolor images are saved

IceFloeTracker.make_filename — Method

make_filename()

Makes default filename with timestamp.

IceFloeTracker.make_psi_s — Method

make_psi_s(XY::Matrix{<:Number};rangeout::Bool=true,
unwrap::Bool=true)

Alternate method of make_psi_s accepting input vectors x and y as a 2-column matrix [x y] in order to facillitate workflow (output from resample_boundary).

IceFloeTracker.make_psi_s — Method

make_psi_s(x::Vector{<:Number},
           y::Vector{<:Number};
           rangeout::Bool=true,
           unwrap::Bool=true)::Tuple{Vector{Float64}, Vector{Float64}}

Builds the ψ-s curve defined by vectors x and y.

Returns a tuple of vectors with the phases ψ in the first component and the traversed arclength in the second component.

Following the convention in [1], the wrapped ψ-s curve has values in [0, 2π) by default; use rangeout to control this behavior.

See also bwtraceboundary, resample_boundary

Arguments

x: Vector of x-coordinates

y: corresponding vector of y-coordinates

rangeout: true (default) for phase values in [0, 2π); false for phase values in (-π, π].

unwrap: set to true to get "unwrapped" phases (default).

Reference

[1] McConnell, Ross, et al. "psi-s correlation and dynamic time warping: two methods for tracking ice floes in SAR images." IEEE Transactions on Geoscience and Remote sensing 29.6 (1991): 1004-1012.

Example

The example below builds a cardioid and obtains its ψ-s curve.

julia> t = range(0,2pi,201);

julia> x = @. cos(t)*(1-cos(t));

julia> y = @. sin(t)*(1-cos(t));

julia> plot(x,y) # visualize the cardioid

julia> psi, s = make_psi_s(x,y);

julia> [s psi] # inspect psi-s data
200×2 Matrix{Float64}:
 0.00049344  0.0314159
 0.0019736   0.0733034
 0.00444011  0.11938
 0.00789238  0.166055
 0.0123296   0.212929
 0.0177505   0.259894
 0.024154    0.306907
 0.0315383   0.35395
 0.0399017   0.401012
 0.0492421   0.448087
 ⋮
 7.96772     9.02377
 7.97511     9.07083
 7.98151     9.11787
 7.98693     9.16488
 7.99137     9.21185
 7.99482     9.25872
 7.99729     9.3054
 7.99877     9.35147
 7.99926     9.39336

 julia> plot(s, psi) # inspect psi-s curve -- should be a straight line from (0, 0) to (2π, 3π)

IceFloeTracker.makecontourstartatP — Method

Make contour start at point P by permuting the elements in contour.

IceFloeTracker.move — Method

move from current pixel to the next in given direction

IceFloeTracker.norm — Method

norm(v)

Get the euclidean norm of the vector v.

IceFloeTracker.normalize_image — Method

normalize_image(truecolor_image, landmask, struct_elem; lambda, kappa, niters, nbins, rblocks, cblocks, clip, smoothing_param, intensity)

Adjusts truecolor land-masked image to highlight ice floe features. This function performs diffusion, adaptive histogram equalization, and sharpening, and returns a greyscale normalized image.

Arguments

truecolor_image: input image in truecolor

landmask: bitmatrix landmask for region of interest

struct_elem: structuring element for dilation

lambda: speed of diffusion (0–0.25)

kappa: conduction coefficient for diffusion (25–100)

niters: number of iterations of diffusion

nbins: number of bins during histogram equalization

rblocks: number of row blocks to divide input image during equalization

cblocks: number of column blocks to divide input image during equalization

clip: threshold for clipping histogram bins (0–1); values closer to one minimize contrast enhancement, values closer to zero maximize contrast enhancement

smoothing_param: pixel radius for gaussian blurring (1–10)

intensity: amount of sharpening to perform

IceFloeTracker.regionprops — Function

regionprops(label_img, ; properties, connectivity)

A wrapper of the regionprops function from the skimage python library.

See its full documentation at https://scikit-image.org/docs/stable/api/skimage.measure.html#skimage.measure.regionprops.

Arguments

label_img: Image with the labeled objects of interest

intensity_img: (Optional) Used for generating extra_properties, integer/float array from which (presumably) label_img was generated

extra_properties: (Optional) not yet implemented. It will be set to nothing

See also regionprops_table

Examples

julia> Random.seed!(123);

julia> bw_img = rand([0, 1], 5, 10)
5×10 Matrix{Int64}:
 1  0  1  0  0  0  0  0  0  1
 1  0  1  1  1  0  0  0  1  1
 1  1  0  1  1  0  1  0  0  1
 0  1  0  1  0  0  0  0  1  0
 1  0  0  0  0  1  0  1  0  1

 julia> label_img = Images.label_components(bw_img, trues(3,3))
 5×10 Matrix{Int64}:
  1  0  1  0  0  0  0  0  0  4
  1  0  1  1  1  0  0  0  4  4
  1  1  0  1  1  0  3  0  0  4
  0  1  0  1  0  0  0  0  4  0
  1  0  0  0  0  2  0  4  0  4

 julia> regions = regionprops(label_img, bw_img);

 julia> for region in regions
           println(region.area,"	", region.perimeter)
        end
13      11.621320343559642
1       0.0
1       0.0
7       4.621320343559642

IceFloeTracker.regionprops_table — Function

regionprops_table(label_img, intensity_img; properties, connectivity, extra_properties, dataframe)

A wrapper of the regionprops_table function from the skimage python library.

See its full documentation at https://scikit-image.org/docs/stable/api/skimage.measure.html#regionprops-table.

Arguments

label_img: Image with the labeled objects of interest

intensity_img: (Optional) Used for generating extra_properties, integer/float array from which (presumably) label_img was generated

properties: List (Vector or Tuple) of properties to be generated for each connected component in label_img

dataframe: (Optional) Bool indicating the output data structure to be a dataframe (with value true). A Dict is returned if dataframe is set to false. Default value: false

extra_properties: (Optional) not yet implemented. It will be set to nothing

See also regionprops

Examples

julia> Random.seed!(123);

julia> bw_img = rand([0, 1], 5, 10)
5×10 Matrix{Int64}:
 1  0  1  0  0  0  0  0  0  1
 1  0  1  1  1  0  0  0  1  1
 1  1  0  1  1  0  1  0  0  1
 0  1  0  1  0  0  0  0  1  0
 1  0  0  0  0  1  0  1  0  1

julia> label_img = Images.label_components(bw_img, trues(3,3))
5×10 Matrix{Int64}:
 1  0  1  0  0  0  0  0  0  4
 1  0  1  1  1  0  0  0  4  4
 1  1  0  1  1  0  3  0  0  4
 0  1  0  1  0  0  0  0  4  0
 1  0  0  0  0  2  0  4  0  4

julia> properties = ["area", "perimeter"]
2-element Vector{String}:
 "area"
 "perimeter"

 julia> regionprops_table(label_img, bw_img, properties = properties, dataframe=true)
 4×2 DataFrame
  Row │ area   perimeter 
      │ Int32  Float64   
 ─────┼──────────────────
    1 │    13   11.6213
    2 │     1    0.0
    3 │     1    0.0
    4 │     7    4.62132

IceFloeTracker.remove_landmask — Method

remove_landmask(landmask, ice_mask)

Find the pixel indexes that are floating ice rather than soft or land ice. Returns an array of pixel indexes.

Arguments

landmask: bitmatrix landmask for region of interest

ice_mask: bitmatrix with ones equal to ice, zeros otherwise

IceFloeTracker.remove_padding — Method

remove_padding(paddedimg, border_spec)

Removes padding from the boundary of padded image paddedimg according to the border specification border_spec type. Returns the cropped image.

Arguments

paddedimg: Pre-padded image.

border_spec: Type representing the style of padding (such as Pad or Fill) with which paddedimg is assumend to be pre-padded. Example: Pad((1,2), (3,4)) specifies 1 row on the top, 2 columns on the left, 3 rows on the bottom, and 4 columns on the right boundary.

See also add_padding

IceFloeTracker.segmentation_A — Method

segmentation_A(reflectance_image, ice_water_discriminated_image, landmask, cloudmask; band_7_threshold, band_2_threshold, band_1_threshold, band_7_relaxed_threshold, band_1_relaxed_threshold, possible_ice_threshold, min_opening_area, fill_range)

Convert a 3-channel false color reflectance image to a 1-channel binary matrix with ice floes contrasted from background. Returns an image segmented and processed. Default thresholds are defined in the published Ice Floe Tracker article: Remote Sensing of the Environment 234 (2019) 111406.

Arguments

reflectance_image: corrected reflectance false color image - bands [7,2,1]

ice_water_discrimination_image: output image from ice-water-discrimination.jl

landmask: bitmatrix landmask for region of interest

cloudmask: bitmatrix cloudmask for region of interest

band_7_threshold: threshold value used to identify ice in band 7, N0f8(RGB intensity/255)

band_2_threshold: threshold value used to identify ice in band 2, N0f8(RGB intensity/255)

band_1_threshold: threshold value used to identify ice in band 2, N0f8(RGB intensity/255)

band_7_relaxed_threshold: threshold value used to identify ice in band 7 if not found on first pass, N0f8(RGB intensity/255)

band_1_relaxed_threshold: threshold value used to identify ice in band 1 if not found on first pass, N0f8(RGB intensity/255)

possible_ice_threshold: threshold value used to identify ice if not found on first or second pass

min_opening_area: minimum size of pixels to use during morphoilogical opening

fill_range: range of values dictating the size of holes to fill

IceFloeTracker.segmentation_B — Method

segmentation_B(sharpened_image, cloudmask, segmented_a_ice_mask, struct_elem; fill_range, isolation_threshold, alpha_level, adjusted_ice_threshold)

Performs image processing and morphological filtering with intermediate files from normalization.jl and segmentation_A to further isolate ice floes, returning a mask of potential ice.

Arguments

sharpened_image: non-cloudmasked but sharpened image, output from normalization.jl

cloudmask: bitmatrix cloudmask for region of interest

segmented_a_ice_mask: binary cloudmasked ice mask from segmentation_a_direct.jl

struct_elem: structuring element for dilation

fill_range: range of values dictating the size of holes to fill

isolation_threshold: threshold used to isolated pixels from sharpened_image; between 0-1

alpha_level: alpha threshold used to adjust contrast

gamma_factor: amount of gamma adjustment

adjusted_ice_threshold: threshold used to set ice equal to one after gamma adjustment

IceFloeTracker.@persist — Macro

@persist img fname or
@persist(img,fname) or
@persist img or
@persist(img)

Given a reference to an image object img, the macro persists (saves to a file) img to the current working directory using fname as filename. Returns img.

Arguments

img: Symbol expression representing an image object loaded in memory.

fname: Optional filename for the persisted image.

Index

IceFloeTracker.add_padding

IceFloeTracker.apply_cloudmask

IceFloeTracker.apply_landmask

IceFloeTracker.atan2

IceFloeTracker.bwareamaxfilt

IceFloeTracker.bwareamaxfilt!

IceFloeTracker.bwperim

IceFloeTracker.bwtraceboundary

IceFloeTracker.check_fname

IceFloeTracker.clockwise

IceFloeTracker.counterclockwise

IceFloeTracker.create_cloudmask

IceFloeTracker.create_landmask

IceFloeTracker.crosscorr

IceFloeTracker.detect_move!

IceFloeTracker.discriminate_ice_water

IceFloeTracker.filt_except_label

IceFloeTracker.filt_except_label!

IceFloeTracker.from_to

IceFloeTracker.get_areas

IceFloeTracker.get_max_label

IceFloeTracker.grad

IceFloeTracker.grad

IceFloeTracker.isincountourlist

IceFloeTracker.landmask

IceFloeTracker.make_filename

IceFloeTracker.make_psi_s

IceFloeTracker.make_psi_s

IceFloeTracker.makecontourstartatP

IceFloeTracker.move

IceFloeTracker.norm

IceFloeTracker.normalize_image

IceFloeTracker.regionprops

IceFloeTracker.regionprops_table

IceFloeTracker.remove_landmask

IceFloeTracker.remove_padding

IceFloeTracker.segmentation_A

IceFloeTracker.segmentation_B

IceFloeTracker.@persist