IceFloeTracker

Documentation for IceFloeTracker.jl package.

IceFloeTracker

Functions

Index

Functions

IceFloeTracker._adjust_histogram — Method

_adjust_histogram(masked_view, nbins, rblocks, cblocks, clip)

Perform adaptive histogram equalization to a masked image. To be invoked within imsharpen.

Arguments

masked_view: input image in truecolor

See imsharpen for a description of the remaining arguments

IceFloeTracker._bin9todec — Method

_bin9todec(v)

Get decimal representation of a bit vector v with the leading bit at its leftmost posistion.

Example

julia> _bin9todec([0 0 0 0 0 0 0 0 0])    
0

julia> _bin9todec([1 1 1 1 1 1 1 1 1])    
511

IceFloeTracker._branch_filter — Function

_branch_filter(
img::AbstractArray{Bool},
func1::Function=branch_candidates_func,
func2::Function=connected_background_count,)

Filter img with _operator_lut using lut1 and lut2.

IceFloeTracker._operator_lut — Method

_operator_lut(I, img, nhood, lut1, lut2)

Look up the neighborhood nhood in lookup tables lut1 and lut2.

Handles cases when the center of nhood is on the edge of img using data in I.

IceFloeTracker.add_padding — Method

add_padding(img, style)

Extrapolate the image img according to the style specifications type. Returns the extrapolated image.

Arguments

img: Image to be padded.

style: A supported type (such as Pad or Fill) representing the extrapolation style. See the relevant documentation for details.

See also remove_padding

IceFloeTracker.apply_cloudmask — Method

apply_cloudmask(reflectance_image, cloudmask)

Zero out pixels containing clouds where clouds and ice are not discernable. Arguments should be of the same size.

Arguments

reference_image: corrected reflectance false color image - bands [7,2,1] or grayscale

cloudmask: binary cloudmask with clouds = 0, else = 1

IceFloeTracker.apply_landmask — Method

apply_landmask(input_image, landmask_binary)

Zero out pixels in land and soft ice regions on truecolor image, return RGB image with zero for all three channels on land/soft ice.

Arguments

input_image: truecolor RGB image

landmask_binary: binary landmask with 1=land, 0=water/ice

IceFloeTracker.atan2 — Method

atan2(y,x)

Wrapper of Base.atan that returns the angle of vector (x,y) in the range [0, 2π).

IceFloeTracker.binarize_landmask — Method

binarize_landmask(landmask_image)

Convert a 3-channel RGB land mask image to a 1-channel binary matrix; land = 0, water/ice = 1.

Arguments

landmask_image: RGB land mask image from fetchdata

IceFloeTracker.branch — Method

branch(img::AbstractArray{Bool})

Find branch points in skeletonized image img according to Definition 3 of [1].

[1] Arcelli, Carlo, and Gabriella Sanniti di Baja. "Skeletons of planar patterns." Machine Intelligence and Pattern Recognition. Vol. 19. North-Holland, 1996. 99-143.

IceFloeTracker.branch_candidates_func — Method

branch_candidates_func(nhood)

Filter nhood as candidate for branch point.

To be passed to the make_lut function.

IceFloeTracker.bridge — Method

bridge(bw)

Set 0-valued pixels to 1 if they have two nonzero neighbors that are not connected. Note the following exceptions:

0 0 0 0 0 0 1 0 1 becomes 1 1 1 0 0 0 0 0 0

1 0 1 1 1 1 0 0 0 becomes 0 0 0 0 0 0 0 0 0

The same applies to all their corresponding rotations.

Examples

julia> bw = [0 0 0; 0 0 0; 1 0 1]
3×3 Matrix{Int64}:
 0  0  0
 0  0  0
 1  0  1

julia> bridge(bw)
3×3 BitMatrix:
 0  0  0
 0  0  0
 1  1  1

julia> bw = [1 0 0; 1 0 1; 0 0 1]
3×3 Matrix{Int64}:
 1  0  0
 1  0  1
 0  0  1

julia> bridge(bw)
3×3 BitMatrix:
 1  1  0
 1  1  1
 0  1  1

 julia> bw = [1 0 1; 0 0 0; 1 0 1]
3×3 Matrix{Int64}:
 1  0  1
 0  0  0
 1  0  1

julia> bridge(bw)
3×3 BitMatrix:
 1  1  1
 1  1  1
 1  1  1

IceFloeTracker.bwareamaxfilt — Function

bwareamaxfilt(bwimg::AbstractArray{Bool}, conn)

Filter the smaller (by area) connected components in bwimg keeping the (assumed unique) largest.

Uses 8-pixel connectivity by default (conn=8). Use conn=4 for 4-pixel connectivity.

IceFloeTracker.bwareamaxfilt! — Function

bwareamaxfilt!(bwimg::AbstractArray)

In-place version of bwareamaxfilt.

See also bwareamaxfilt

IceFloeTracker.bwdist — Method

bwdist(bwimg)

Distance transform for binary image bwdist.

IceFloeTracker.bwperim — Method

bwperim(bwimg)

Locate the pixels at the boundary of objects in an binary image bwimg using 8-pixel connectivity.

Arguments

bwimg: Binary (black/white – 1/0) image

Examples

julia> A = zeros(Bool, 13, 16); A[2:6, 2:6] .= 1; A[4:8, 7:10] .= 1; A[10:12,13:15] .= 1; A[10:12,3:6] .= 1;

julia> A
13×16 Matrix{Bool}:
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0

 julia> bwperim(A)
13×16 Matrix{Bool}:
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  0  0  0  1  0  0  0  0  0  0  0  0  0  0
 0  1  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  1  0  0  0  0  0  0  0  1  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  0  0  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  0  0  1  0  0  0  0  0  0  1  0  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0

IceFloeTracker.bwtraceboundary — Method

bwtraceboundary(image::Union{Matrix{Int64},Matrix{Float64},BitMatrix};
                P0::Union{Tuple{Int,Int},CartesianIndex{2},Nothing}=nothing,
                closed::Bool=true)

Trace the boundary of objects in image

Background pixels are represented as zero. The algorithm traces the boundary counterclockwise and an initial point P0 can be specified. If more than one boundary is detected and an initial point is provided, the boundary that contains this point is returned as a vector of CartesianIndex types. Otherwise an array of vectors is returned with all the detected boundaries in image.

Arguments

image: image, preferably binary with one single object, whose objects' boundaries are to be traced.

P0: initial point of a target boundary.

closed: if true (default) makes the inital point of a boundary equal to the last point.

Example

julia> A = zeros(Int, 13, 16); A[2:6, 2:6] .= 1; A[4:8, 7:10] .= 1; A[10:12,13:15] .= 1; A[10:12,3:6] .= 1;

julia> A
13×16 Matrix{Int64}:
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  1  1  1  1  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  1  1  1  1  0  0  0  0  0  0  1  1  1  0
 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0

julia> boundary = IceFloeTracker.bwtraceboundary(A);

julia> boundary[3]
9-element Vector{CartesianIndex}:
 CartesianIndex(10, 13)
 CartesianIndex(11, 13)
 CartesianIndex(12, 13)
 CartesianIndex(12, 14)
 CartesianIndex(12, 15)
 CartesianIndex(11, 15)
 CartesianIndex(10, 15)
 CartesianIndex(10, 14)
 CartesianIndex(10, 13)

IceFloeTracker.cache_vector — Method

cache_vector(type::Type, numel::Int64, size::Tuple{Int64, Int64})::Vector{type}

Build a vector of types type with numel elements of size size.

Example

julia> cache_vector(Matrix{Float64}, 3, (2, 2))
3-element Vector{Matrix{Float64}}:
 [0.0 6.9525705991269e-310; 6.9525705991269e-310 0.0]
 [0.0 6.9525705991269e-310; 6.9525705991269e-310 0.0]
 [0.0 6.95257028858726e-310; 6.95257029000147e-310 0.0]

IceFloeTracker.check_fname — Function

check_fname(fname)

Checks fname does not exist in current directory; throws an assertion if this condition is false.

Arguments

fname: String object or Symbol to a reference to a String representing a path.

IceFloeTracker.clockwise — Method

Make a clockwise turn from the dir direction

IceFloeTracker.connected_background_count — Method

connected_background_count(nhood)

Second lut generator for neighbor transform with diamond strel (4-neighborhood).

To be passed to the make_lut function.

IceFloeTracker.counterclockwise — Method

Make a counterclockwise turn from the dir direction

IceFloeTracker.create_cloudmask — Method

create_cloudmask(reflectance_image; prelim_threshold, band_7_threshold, band_2_threshold, ratio_lower, ratio_upper)

Convert a 3-channel false color reflectance image to a 1-channel binary matrix; clouds = 0, else = 1. Default thresholds are defined in the published Ice Floe Tracker article: Remote Sensing of the Environment 234 (2019) 111406.

Arguments

reflectance_image: corrected reflectance false color image - bands [7,2,1]

prelim_threshold: threshold value used to identify clouds in band 7, N0f8(RGB intensity/255)

band_7_threshold: threshold value used to identify cloud-ice in band 7, N0f8(RGB intensity/255)

band_2_threshold: threshold value used to identify cloud-ice in band 2, N0f8(RGB intensity/255)

ratio_lower: threshold value used to set lower ratio of cloud-ice in bands 7 and 2

ratio_upper: threshold value used to set upper ratio of cloud-ice in bands 7 and 2

IceFloeTracker.create_landmask — Method

create_landmask(landmask_image, struct_elem, fill_value_lower, fill_value_upper)

Convert a 3-channel RGB land mask image to a 1-channel binary matrix, including a buffer to extend the land over any soft ice regions; land = 0, water/ice = 1.

Arguments

landmask_image: RGB land mask image from fetchdata

struct_elem: structuring element for dilation (optional)

fill_value_lower: fill holes having at least these many pixels (optional)

fill_value_upper: fill holes having at most these many pixels (optional)

IceFloeTracker.crosscorr — Method

r, lags = crosscorr(u::Vector{T},
                    v::Vector{T};
                    normalize::Bool=false,
                    padmode::Symbol=:longest)

Wrapper of DSP.xcorr with normalization (see https://docs.juliadsp.org/stable/convolutions/#DSP.xcorr)

Returns the pair (r, lags) with the cross correlation scores r and corresponding lags according to padmode.

Arguments

u,v: real vectors which could have unequal length.

normalize: return normalized correlation scores (false by default).

padmode: either :longest (default) or :none to control padding of shorter vector with zeros.

Examples

The example below builds two vectors, one a shifted version of the other, and computes various cross correlation scores.

julia> n = 1:5;

julia> x = 0.48.^n;

julia> y = circshift(x,3);

julia> r, lags = crosscorr(x,y,normalize=true);

julia> [r lags]
9×2 Matrix{Float64}:
0.369648    -4.0
0.947531    -3.0
0.495695    -2.0
0.3231      -1.0
0.332519     0.0
0.15019      1.0
0.052469     2.0
0.0241435    3.0
0.00941878   4.0

julia> r, lags = crosscorr(x,y,normalize=true,padmode=:none);

julia> [r lags]
9×2 Matrix{Float64}:
0.369648    1.0
0.947531    2.0
0.495695    3.0
0.3231      4.0
0.332519    5.0
0.15019     6.0
0.052469    7.0
0.0241435   8.0
0.00941878  9.0

This final example builds two vectors of different length and computes some cross correlation scores.

julia> n = 1:5; m = 1:3;

julia> x = 0.48.^n; y = 0.48.^m;

julia> r, lags = crosscorr(x,y,normalize=true);

julia> [r lags]
9×2 Matrix{Float64}:
0.0          -4.0
4.14728e-17  -3.0
0.178468     -2.0
0.457473     -1.0
0.994189      0.0
0.477211      1.0
0.229061      2.0
0.105402      3.0
0.0411191     4.0

julia> r, lags = crosscorr(x,y,normalize=true,padmode=:none);

julia> [r lags]
7×2 Matrix{Float64}:
0.178468   1.0
0.457473   2.0
0.994189   3.0
0.477211   4.0
0.229061   5.0
0.105402   6.0
0.0411191  7.0

IceFloeTracker.detect_move! — Method

Workhorse function: Get all pixel coords for detected border.

IceFloeTracker.disc_ice_water — Method

disc_ice_water(
reflectance_imgs::Vector{Matrix{RGB{Float64}}},
sharpened_imgs::Vector{Matrix{Float64}},
cloudmasks::Vector{BitMatrix},
landmask::BitMatrix,

)

Generate vector of ice/water discriminated images from the collection of reflectance, sharpened truecolor, and cloudmask images using the study area landmask. Returns a vector of ice/water masks.

IceFloeTracker.discriminate_ice_water — Method

discriminate_ice_water(
reflectance_image::Matrix{RGB{Float64}}, 
image_sharpened::Matrix{Float64}, 
landmask_bitmatrix::T,
cloudmask_bitmatrix::T,
floes_threshold::Float64=Float64(100 / 255),
mask_clouds_lower::Float64=Float64(17 / 255),
mask_clouds_upper::Float64=Float64(30 / 255),
kurt_thresh_lower::Real=2,
kurt_thresh_upper::Real=8,
skew_thresh::Real=4,
st_dev_thresh_lower::Float64=Float64(84 / 255),
st_dev_thresh_upper::Float64=Float64(98.9 / 255),
clouds_ratio_threshold::Float64=0.02,
differ_threshold::Float64=0.6,
nbins::Real=155

)::AbstractMatrix where T<:AbstractArray{Bool}

Generates an image with ice floes apparent after filtering and combining previously processed versions of reflectance and truecolor images from the same region of interest. Returns an image ready for segmentation to isolate floes.

Arguments

reflectance_image: input image in false color reflectance

image_sharpened: sharpenened version of true color image

landmask_bitmatrix: landmask for region of interest

cloudmask_bitmatrix: cloudmask for region of interest

floes_threshold: heuristic applied to original reflectance image

mask_clouds_lower: lower heuristic applied to mask out clouds

mask_clouds_upper: upper heuristic applied to mask out clouds

kurt_thresh_lower: lower heuristic used to set pixel value threshold based on kurtosis in histogram

kurt_thresh_upper: upper heuristic used to set pixel value threshold based on kurtosis in histogram

skew_thresh: heuristic used to set pixel value threshold based on skewness in histogram

st_dev_thresh_lower: lower heuristic used to set pixel value threshold based on standard deviation in histogram

st_dev_thresh_upper: upper heuristic used to set pixel value threshold based on standard deviation in histogram

clouds2_threshold: heuristic used to set pixel value threshold based on ratio of clouds

differ_threshold: heuristic used to calculate proportional intensity in histogram

nbins: number of bins during histogram build

IceFloeTracker.extractfeatures — Method

extractfeatures(bw::AbstractArray{Bool}, area_threshold::Tuple{Int64,Int64}, features::Union{Vector{Symbol},Vector{String}})::DataFrame

Extract features from the labeled image bw using the area thresholds in area_threshold and the vector of features features. Returns a DataFrame of the extracted features.

Arguments

bw`: A labeled image.

min_area and max_area: The minimum and maximum (inclusive) areas of the floes to extract.

features: A vector of the features to extract.

Example

julia> Random.seed!(123);

julia> bw_img = rand(Bool, 5, 50)
5×50 Matrix{Bool}:
 0  0  1  0  1  0  1  0  1  0  1  0  0  0  0  0  0  1  1  0  1  0  1  1  1  1  1  0  0  0  0  0  0  0  1  1  1  0  0  1  1  1  0  0  1  1  1  1  0  1
 1  0  0  0  1  1  0  0  0  1  0  0  0  0  0  0  0  0  0  0  1  1  0  0  1  0  0  0  0  1  1  1  1  1  0  1  0  1  1  0  1  1  1  1  0  0  1  1  1  1
 0  1  1  1  1  1  0  0  1  0  1  1  0  1  1  1  1  1  1  1  1  0  0  1  0  1  1  1  0  0  0  0  1  0  0  0  0  0  1  0  0  0  1  0  1  0  0  0  0  0
 0  1  0  0  1  1  1  0  0  0  1  0  1  0  1  0  1  0  0  1  1  1  1  0  0  1  0  0  0  0  1  0  0  0  1  1  1  0  0  1  1  1  0  1  0  0  0  0  1  0
 0  1  1  1  1  0  0  0  1  0  1  0  0  1  0  1  0  0  1  1  0  1  1  0  1  1  0  1  1  0  0  0  1  0  0  0  0  1  0  0  0  1  1  0  1  0  1  1  1  0

julia> features = ["centroid", "area", "major_axis_length", "minor_axis_length", "convex_area", "bbox"]
6-element Vector{String}:
 "centroid"
 "area"
 "major_axis_length"
 "minor_axis_length"
 "convex_area"
 "bbox"

 julia> min_area, max_area = 1, 5
 (1, 5)

 julia> IceFloeTracker.extractfeatures(bw_img, min_area=min_area, max_area=max_area, features=features)
 8×10 DataFrame
  Row │ area   bbox-0  bbox-1  bbox-2  bbox-3  centroid-0  centroid-1  convex_area  major_axis_length  minor_axis_length 
      │ Int32  Int32   Int32   Int32   Int32   Float64     Float64     Int32        Float64            Float64
 ─────┼──────────────────────────────────────────────────────────────────────────────────────────────────────────────────
    1 │     1       1       3       2       4        0.0         2.0             1            0.0                0.0
    2 │     1       5       9       6      10        4.0         8.0             1            0.0                0.0
    3 │     2       1      18       2      20        0.0        17.5             2            2.0                0.0
    4 │     2       5      28       6      30        4.0        27.5             2            2.0                0.0
    5 │     1       4      31       5      32        3.0        30.0             1            0.0                0.0
    6 │     1       5      33       6      34        4.0        32.0             1            0.0                0.0
    7 │     4       4      35       6      39        3.25       35.5             5            4.68021            1.04674
    8 │     4       4      47       6      50        3.75       47.25            5            3.4641             1.41421
 

IceFloeTracker.filt_except_label! — Method

filt_except_label!(labeled_arr::Array{Int64, 2}, label::Int64)

In-place version of filt_except_label.

See also filt_except_label

IceFloeTracker.filt_except_label — Method

filt_except_label(labeled_arr::Array{Int64, 2}, label::Int64)

Make 0 all values in labeled_arr that are not equal to label.

See also filt_except_label!

IceFloeTracker.find_ice_labels — Method

find_ice_labels(reflectance_image, landmask; band_7_threshold, band_2_threshold, band_1_threshold, band_7_relaxed_threshold, band_1_relaxed_threshold, possible_ice_threshold)

Locate the pixels of obvious ice from false color reflectance image. Returns a binary mask with ice floes contrasted from background. Default thresholds are defined in the published Ice Floe Tracker article: Remote Sensing of the Environment 234 (2019) 111406.

Arguments

reflectance_image: corrected reflectance false color image - bands [7,2,1]

landmask: bitmatrix landmask for region of interest

band_7_threshold: threshold value used to identify ice in band 7, N0f8(RGB intensity/255)

band_2_threshold: threshold value used to identify ice in band 2, N0f8(RGB intensity/255)

band_1_threshold: threshold value used to identify ice in band 2, N0f8(RGB intensity/255)

band_7_relaxed_threshold: threshold value used to identify ice in band 7 if not found on first pass, N0f8(RGB intensity/255)

band_1_relaxed_threshold: threshold value used to identify ice in band 1 if not found on first pass, N0f8(RGB intensity/255)

IceFloeTracker.find_reflectance_peaks — Method

find_reflectance_peaks(reflectance_channel, possible_ice_threshold;)

Find histogram peaks in single channels of a reflectance image and return the second greatest peak. If needed, edges can be returned as the first object from build_histogram. Similarly, peak values can be returned as the second object from findmaxima.

Arguments

reflectance_channel: either band 2 or band 1 of false-color reflectance image

possible_ice_threshold: threshold value used to identify ice if not found on first or second pass

IceFloeTracker.fname_ext_splice — Method

fname_ext_splice(fname, ext)

Join "fname" and "ext" with '.'.

IceFloeTracker.fname_ext_split — Method

fname_ext_split(fname)

Split "fname.ext" into "fname" and "ext".

IceFloeTracker.from_to — Method

Get index in Moore neigborhood representing the direction from the from pixel coords to the to pixel coords (see definition of dir_delta below).

Clockwise Moore neighborhood.

dir_delta = [CartesianIndex(-1, 0), CartesianIndex(-1, 1), CartesianIndex(0, 1), CartesianIndex(1, 1), CartesianIndex(1, 0), CartesianIndex(1, -1), CartesianIndex(0, -1), CartesianIndex(-1,-1)]

IceFloeTracker.get_areas — Method

get_areas(labeled_arr::Array{T, 2})::Dict{T, Int} where T

Get the "areas" (count of pixels of a given label) of the connected components in labeled_arr.

Return a dictionary with the frequency distribution: label => countoflabel.

IceFloeTracker.get_max_label — Method

get_max_label(d::Dict{Int, Int})

Get the label k in dictionary d for which d[k] is maximal.

IceFloeTracker.grad — Method

dx, dy = grad(A::Matrix{<:Number})

Make gradient vector field for the set of points with coordinates in the rows of the matrix A with x-coordinates down column 1 and y-coordinates down column 2. Return a tuple with dx and dy in that order.

IceFloeTracker.grad — Method

dx, dy = grad(x::Vector{<:Number}, y::Vector{<:Number})

Make gradient vector field for the set of points with coordinates in vectors x and y. Return a tuple with dx and dy in that order.

IceFloeTracker.hbreak! — Method

hbreak!(img::AbstractArray{Bool})

Inplace version of hbreak. See also hbreak.

IceFloeTracker.hbreak — Method

hbreak(img::AbstractArray{Bool})

Remove H-connected pixels in the binary image img. See also hbreak! for an inplace version of this function.

Examples

```jldoctest; setup = :(using IceFloeTracker)

julia> h1 = trues(3,3); h1[[1 3], 2] .= false; h1 3×3 BitMatrix: 1 0 1 1 1 1 1 0 1

julia> h2 = trues(3,3); h2[2, [1 3]] .= false; h2 3×3 BitMatrix: 1 1 1 0 1 0 1 1 1

julia> hbreak!(h1); h1 # modify h1 inplace 3×3 BitMatrix: 1 0 1 1 0 1 1 0 1

julia> hbreak(h2) 3×3 BitMatrix: 1 1 1 0 0 0 1 1 1

IceFloeTracker.imextendedmin — Method

imextendedmin(img)

Mimics MATLAB's imextendedmin function that computes the extended-minima transform, which is the regional minima of the H-minima transform. Regional minima are connected components of pixels with a constant intensity value. This function returns a transformed bitmatrix.

Arguments

img: image object

h: suppress minima below this depth threshold

conn: neighborhood connectivity; in 2D 1 = 4-neighborhood and 2 = 8-neighborhood

IceFloeTracker.imsharpen — Function

imsharpen(truecolor_image, lambda, kappa, niters, nbins, rblocks, cblocks, clip, smoothing_param, intensity)

Sharpen truecolor_image.

Arguments

truecolor_image: input image in truecolor

lambda: speed of diffusion (0–0.25)

kappa: conduction coefficient for diffusion (25–100)

niters: number of iterations of diffusion

nbins: number of bins during histogram equalization

rblocks: number of row blocks to divide input image during equalization

cblocks: number of column blocks to divide input image during equalization

clip: Thresholds for clipping histogram bins (0–1); values closer to one minimize contrast enhancement, values closer to zero maximize contrast enhancement

smoothing_param: pixel radius for gaussian blurring (1–10)

intensity: amount of sharpening to perform

IceFloeTracker.imsharpen_gray — Method

imsharpen_gray(imgsharpened, landmask)

Apply landmask and return Gray type image in colorview for normalization.

IceFloeTracker.isincountourlist — Method

Check P is in countour list if so return the index of the contour that contains P, otherwise return false.

IceFloeTracker.kmeans_segmentation — Method

kmeans_segmentation(gray_image, ice_labels;)

Apply k-means segmentation to a gray image to isolate a cluster group representing sea ice. Returns a binary image with ice segmented from background.

Arguments

gray_image: output image from ice-water-discrimination.jl or gray ice floe leads image in segmentation_f.jl

ice_labels: vector if pixel coordinates output from find_ice_labels.jl

IceFloeTracker.landmask — Method

landmask(landmask_raw::Matrix{RGB{N0f8}}; dilate::Bool)

Usage: landmask(landmask_raw, dilate=false).

To build a dilated landmask use landmask(landmask_raw, dilate=true).

IceFloeTracker.landmask — Method

landmask(; input, output)

Given an input directory with a landmask file and possibly truecolor images, create a land/soft ice mask. The resulting images are saved to the snakemake output directory. Returns the landmkask object.

Arguments

input: path to image dir containing truecolor and landmask source images

output: path to output dir where land-masked truecolor images and the generated binary land mask are saved

IceFloeTracker.load_imgs — Method

load_imgs(; input::String, image_type::String)

Load all images of type image_type (either "truecolor" or "reflectance") in input into a vector.

IceFloeTracker.make_filename — Method

make_filename()

Makes default filename with timestamp.

IceFloeTracker.make_hbreak_dict — Method

make_hbreak_dict()

Build dict with the two versions of an H-connected 3x3 neighboorhood.

h1 = [1 0 1 1 1 1 1 0 1]

h2 = [1 1 1 0 1 0 1 1 1]

IceFloeTracker.make_lut — Method

make_lut(lutfunc::Function)

Generate lookup table (lut) for 3x3 neighborhoods according to lutfunc.

IceFloeTracker.make_psi_s — Method

make_psi_s(XY::Matrix{<:Number};rangeout::Bool=true,
unwrap::Bool=true)

Alternate method of make_psi_s accepting input vectors x and y as a 2-column matrix [x y] in order to facillitate workflow (output from resample_boundary).

IceFloeTracker.make_psi_s — Method

make_psi_s(x::Vector{<:Number},
           y::Vector{<:Number};
           rangeout::Bool=true,
           unwrap::Bool=true)::Tuple{Vector{Float64}, Vector{Float64}}

Builds the ψ-s curve defined by vectors x and y.

Returns a tuple of vectors with the phases ψ in the first component and the traversed arclength in the second component.

Following the convention in [1], the wrapped ψ-s curve has values in [0, 2π) by default; use rangeout to control this behavior.

See also bwtraceboundary, resample_boundary

Arguments

x: Vector of x-coordinates

y: corresponding vector of y-coordinates

rangeout: true (default) for phase values in [0, 2π); false for phase values in (-π, π].

unwrap: set to true to get "unwrapped" phases (default).

Reference

[1] McConnell, Ross, et al. "psi-s correlation and dynamic time warping: two methods for tracking ice floes in SAR images." IEEE Transactions on Geoscience and Remote sensing 29.6 (1991): 1004-1012.

Example

The example below builds a cardioid and obtains its ψ-s curve.

julia> t = range(0,2pi,201);

julia> x = @. cos(t)*(1-cos(t));

julia> y = @. sin(t)*(1-cos(t));

julia> plot(x,y) # visualize the cardioid

julia> psi, s = make_psi_s(x,y);

julia> [s psi] # inspect psi-s data
200×2 Matrix{Float64}:
 0.00049344  0.0314159
 0.0019736   0.0733034
 0.00444011  0.11938
 0.00789238  0.166055
 0.0123296   0.212929
 0.0177505   0.259894
 0.024154    0.306907
 0.0315383   0.35395
 0.0399017   0.401012
 0.0492421   0.448087
 ⋮
 7.96772     9.02377
 7.97511     9.07083
 7.98151     9.11787
 7.98693     9.16488
 7.99137     9.21185
 7.99482     9.25872
 7.99729     9.3054
 7.99877     9.35147
 7.99926     9.39336

 julia> plot(s, psi) # inspect psi-s curve -- should be a straight line from (0, 0) to (8, 3π)

IceFloeTracker.makecontourstartatP — Method

Make contour start at point P by permuting the elements in contour.

IceFloeTracker.mismatch — Function

mismatch(fixed::AbstractArray,
              moving::AbstractArray,
              mxshift::Tuple{Int64, Int64}=(10,10),
              mxrot::Float64=pi/4;
              kwargs...
              )

Estimate a rigid transformation (translation + rotation) that minimizes the 'mismatch' of aligning moving with fixed using the QuadDIRECT algorithm.

Returns a pair with the mismatch score mm and the associated transformation tfm.

Arguments

fixed,moving: images to align via a rigid transformation

mxshift: maximum allowed translation in units of array indices (default set to (10,10))

mxrot: maximum allowed rotation in radians (default set to π/4)

thresh: minimum sum-of-squared-intensity overlap between the images (default is 10% of the sum-of-squared-intensity of fixed)

kwargs: other arguments such as tol, ftol, and fvalue (see QuadDIRECT.analyze for details)

```

IceFloeTracker.move — Method

move from current pixel to the next in given direction

IceFloeTracker.norm — Method

norm(v)

Get the euclidean norm of the vector v.

IceFloeTracker.normalize_image — Method

normalize_image(image_sharpened, image_sharpened_gray, landmask, struct_elem;)

Adjusts sharpened land-masked image to highlight ice floe features.

Does reconstruction and landmasking to image_sharpened.

Arguments

image_sharpened: sharpened image (output of imsharpen)

image_sharpened_gray: grayscale, landmasked sharpened image (output of imsharpen_gray(image_sharpened))

landmask: landmask for region of interest

struct_elem: structuring element for dilation

IceFloeTracker.padnhood — Method

padnhood(img, I, nhood)

Pad the matrix img[nhood] with zeros according to the position of I within the edgesimg.

Returns img[nhood] if I is not an edge index.

IceFloeTracker.prune — Method

prune(img::AbstractArray{Bool})

Remove foreground pixels in the binary image img with degree less than 3 (fewer than 3 linear or diagonal neighboring 1s).

Examples

```jldoctest; setup = :(using IceFloeTracker)

julia> bw = zeros(Bool, 15, 15);

julia> bw[:,8] .= true; bw[8,:] .= true; bw[7:9, 7:9] .= true;

julia> [bw[i,i] = 1 for i=1:15]; # fill diagonal

julia> [bw[15-i,i+1] = 1 for i=0:14]; # fill the other diagonal

julia> bw 15×15 Matrix{Bool}: 1 0 0 0 0 0 0 1 0 0 0 0 0 0 1 0 1 0 0 0 0 0 1 0 0 0 0 0 1 0 0 0 1 0 0 0 0 1 0 0 0 0 1 0 0 0 0 0 1 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0 1 0 0 1 0 0 1 0 0 0 0 0 0 0 0 0 1 0 1 0 1 0 0 0 0 0 0 0 0 0 0 0 1 1 1 0 0 0 0 0 0 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 0 0 0 0 0 0 1 1 1 0 0 0 0 0 0 0 0 0 0 0 1 0 1 0 1 0 0 0 0 0 0 0 0 0 1 0 0 1 0 0 1 0 0 0 0 0 0 0 1 0 0 0 1 0 0 0 1 0 0 0 0 0 1 0 0 0 0 1 0 0 0 0 1 0 0 0 1 0 0 0 0 0 1 0 0 0 0 0 1 0 1 0 0 0 0 0 0 1 0 0 0 0 0 0 1

julia> prune(bw) 15×15 Matrix{Bool}: 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 1 0 0 0 0 0 0 0 0 0 0 0 1 1 1 1 1 0 0 0 0 0 0 0 0 0 0 0 1 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0

IceFloeTracker.regionprops — Function

regionprops(label_img, ; properties, connectivity)

A wrapper of the regionprops function from the skimage python library.

See its full documentation at https://scikit-image.org/docs/stable/api/skimage.measure.html#skimage.measure.regionprops.

Arguments

label_img: Image with the labeled objects of interest

intensity_img: (Optional) Used for generating extra_properties, integer/float array from which (presumably) label_img was generated

extra_properties: (Optional) not yet implemented. It will be set to nothing

See also regionprops_table

Examples

julia> Random.seed!(123);

julia> bw_img = rand([0, 1], 5, 10)
5×10 Matrix{Int64}:
 1  0  1  0  0  0  0  0  0  1
 1  0  1  1  1  0  0  0  1  1
 1  1  0  1  1  0  1  0  0  1
 0  1  0  1  0  0  0  0  1  0
 1  0  0  0  0  1  0  1  0  1

 julia> label_img = Images.label_components(bw_img, trues(3,3))
 5×10 Matrix{Int64}:
  1  0  1  0  0  0  0  0  0  4
  1  0  1  1  1  0  0  0  4  4
  1  1  0  1  1  0  3  0  0  4
  0  1  0  1  0  0  0  0  4  0
  1  0  0  0  0  2  0  4  0  4

 julia> regions = regionprops(label_img, bw_img);

 julia> for region in regions
           println(region.area,"	", region.perimeter)
        end
13      11.621320343559642
1       0.0
1       0.0
7       4.621320343559642

IceFloeTracker.regionprops_table — Function

regionprops_table(label_img, intensity_img; properties, connectivity, extra_properties)

A wrapper of the regionprops_table function from the skimage python library.

See its full documentation at https://scikit-image.org/docs/stable/api/skimage.measure.html#regionprops-table.

Arguments

label_img: Image with the labeled objects of interest

intensity_img: (Optional) Used for generating extra_properties, integer/float array from which (presumably) label_img was generated

properties: List (Vector or Tuple) of properties to be generated for each connected component in label_img

extra_properties: (Optional) not yet implemented. It will be set to nothing

See also regionprops

Examples

julia> Random.seed!(123);

julia> bw_img = rand([0, 1], 5, 10)
5×10 Matrix{Int64}:
 1  0  1  0  0  0  0  0  0  1
 1  0  1  1  1  0  0  0  1  1
 1  1  0  1  1  0  1  0  0  1
 0  1  0  1  0  0  0  0  1  0
 1  0  0  0  0  1  0  1  0  1

julia> label_img = Images.label_components(bw_img, trues(3,3))
5×10 Matrix{Int64}:
 1  0  1  0  0  0  0  0  0  4
 1  0  1  1  1  0  0  0  4  4
 1  1  0  1  1  0  3  0  0  4
 0  1  0  1  0  0  0  0  4  0
 1  0  0  0  0  2  0  4  0  4

julia> properties = ["area", "perimeter"]
2-element Vector{String}:
 "area"
 "perimeter"

 julia> regionprops_table(label_img, bw_img, properties = properties, dataframe=true)
 4×2 DataFrame
  Row │ area   perimeter 
      │ Int32  Float64   
 ─────┼──────────────────
    1 │    13   11.6213
    2 │     1    0.0
    3 │     1    0.0
    4 │     7    4.62132

IceFloeTracker.remove_padding — Method

remove_padding(paddedimg, border_spec)

Removes padding from the boundary of padded image paddedimg according to the border specification border_spec type. Returns the cropped image.

Arguments

paddedimg: Pre-padded image.

border_spec: Type representing the style of padding (such as Pad or Fill) with which paddedimg is assumend to be pre-padded. Example: Pad((1,2), (3,4)) specifies 1 row on the top, 2 columns on the left, 3 rows on the bottom, and 4 columns on the right boundary.

See also add_padding

IceFloeTracker.segmentation_A — Method

segmentation_A(segmented_ice_cloudmasked; min_opening_area)

Apply k-means segmentation to a gray image to isolate a cluster group representing sea ice. Returns an image segmented and processed as well as an intermediate files needed for downstream functions.

Arguments

segmented_ice_cloudmask: bitmatrix with open water/clouds = 0, ice = 1, output from segmented_ice_cloudmasking()

min_opening_area: minimum size of pixels to use during morphological opening

fill_range: range of values dictating the size of holes to fill

IceFloeTracker.segmentation_B — Function

segmentation_B(sharpened_image, cloudmask, segmented_a_ice_mask, struct_elem; fill_range, isolation_threshold, alpha_level, adjusted_ice_threshold)

Performs image processing and morphological filtering with intermediate files from normalization.jl and segmentation_A to further isolate ice floes, returning a mask of potential ice.

Arguments

sharpened_image: non-cloudmasked but sharpened image, output from normalization.jl

cloudmask: bitmatrix cloudmask for region of interest

segmented_a_ice_mask: binary cloudmasked ice mask from segmentation_a_direct.jl

struct_elem: structuring element for dilation

fill_range: range of values dictating the size of holes to fill

isolation_threshold: threshold used to isolated pixels from sharpened_image; between 0-1

alpha_level: alpha threshold used to adjust contrast

gamma_factor: amount of gamma adjustment

adjusted_ice_threshold: threshold used to set ice equal to one after gamma adjustment

IceFloeTracker.segmentation_D_E — Method

segmentation_D_E(watershed_D, watershed_E;)

Intersects the outputs of watershed segmentation on intermediate files from segmentation B and C, indicating potential sparse boundaries of ice floes.

Arguments

watershed_D: binary segmentation mask from segmentation_D

watershed_E: binary segmentation mask from segmentation_E

IceFloeTracker.segmentation_F — Method

segmentation_F(
segmentation_C_ice_mask::BitMatrix,
segmentation_B_not_ice_mask::BitMatrix,
watershed_intersect::BitMatrix,
cloudmask::BitMatrix,
landmask::BitMatrix,
ice_labels::Vector{Int64};
min_area_opening::Int64=20,

)::BitMatrix

Cleans up past segmentation images with morphological operations, and applies the results of prior watershed segmentation, returning the final cleaned image for tracking with ice floes segmented and isolated.

Arguments

segmentation_C_ice_mask: binary cloudmasked, landmasked intermediate file from segmentationC (`segmentedice`)

segmentation_B_not_ice_mask: binary mask output from segmentation_b.jl

watershed_intersect: ice pixels, output from segmentation_d_e.jl

cloudmask.jl: bitmatrix cloudmask for region of interest

landmask.jl: bitmatrix landmask for region of interest

ice_labels: vector of pixel coordinates output from find_ice_labels.jl

min_area_opening: threshold used for area opening; pixel groups greater than threshold are retained

IceFloeTracker.segmented_ice_cloudmasking — Method

segmented_ice_cloudmasking(gray_image, cloudmask, ice_labels;)

Apply cloudmask to a bitmatrix of segmented ice after kmeans clustering. Returns a bitmatrix with open water/clouds = 0, ice = 1).

Arguments

gray_image: output image from ice-water-discrimination.jl or gray ice floe leads image in segmentation_f.jl

cloudmask: bitmatrix cloudmask for region of interest

ice_labels: vector if pixel coordinates output from find_ice_labels.jl

IceFloeTracker.sharpen — Method

sharpen(truecolor_imgs::Vector{Matrix{Float64}}, landmask_no_dilate::Matrix{Bool})

Sharpen truecolor images with the non-dilated landmask applied. Returns a vector of sharpened images.

IceFloeTracker.sharpen_gray — Method

sharpen_gray(
sharpened_imgs::Vector{Matrix{Float64}},
landmask::AbstractArray{Bool},

)

Apply the landmask to the collection of sharpened truecolor images and return a gray colorview of the collection.

IceFloeTracker.timestamp — Method

timestamp(fname)

Attach timestamp to fname.

IceFloeTracker.watershed_ice_floes — Method

watershed_ice_floes(intermediate_segmentation_image;)

Performs image processing and watershed segmentation with intermediate files from segmentationb.jl or segmentationc.jl to further isolate ice floes, returning a binary segmentation mask indicating potential sparse boundaries of ice floes.

Arguments

-intermediate_segmentation_image: binary cloudmasked and landmasked intermediate file from segmentation B or C, typically either segmentation_b_not_ice_mask or segmented_c

IceFloeTracker.@persist — Macro

@persist img fname
@persist(img,fname)
@persist img
@persist(img)
@persist img fname ts
@persist(img, fname, ts)

Given a reference to an image object img, the macro persists (saves to a file) img to the current working directory using fname as filename. Returns img.

Arguments

img: Symbol expression representing an image object loaded in memory.

fname: Optional filename for the persisted image.

ts: Optional boolean to attach timestamp to fname.

Index

IceFloeTracker._adjust_histogram

IceFloeTracker._bin9todec

IceFloeTracker._branch_filter

IceFloeTracker._operator_lut

IceFloeTracker.add_padding

IceFloeTracker.apply_cloudmask

IceFloeTracker.apply_landmask

IceFloeTracker.atan2

IceFloeTracker.binarize_landmask

IceFloeTracker.branch

IceFloeTracker.branch_candidates_func

IceFloeTracker.bridge

IceFloeTracker.bwareamaxfilt

IceFloeTracker.bwareamaxfilt!

IceFloeTracker.bwdist

IceFloeTracker.bwperim

IceFloeTracker.bwtraceboundary

IceFloeTracker.cache_vector

IceFloeTracker.check_fname

IceFloeTracker.clockwise

IceFloeTracker.connected_background_count

IceFloeTracker.counterclockwise

IceFloeTracker.create_cloudmask

IceFloeTracker.create_landmask

IceFloeTracker.crosscorr

IceFloeTracker.detect_move!

IceFloeTracker.disc_ice_water

IceFloeTracker.discriminate_ice_water

IceFloeTracker.extractfeatures

IceFloeTracker.filt_except_label

IceFloeTracker.filt_except_label!

IceFloeTracker.find_ice_labels

IceFloeTracker.find_reflectance_peaks

IceFloeTracker.fname_ext_splice

IceFloeTracker.fname_ext_split

IceFloeTracker.from_to

IceFloeTracker.get_areas

IceFloeTracker.get_max_label

IceFloeTracker.grad

IceFloeTracker.grad

IceFloeTracker.hbreak

IceFloeTracker.hbreak!

IceFloeTracker.imextendedmin

IceFloeTracker.imsharpen

IceFloeTracker.imsharpen_gray

IceFloeTracker.isincountourlist

IceFloeTracker.kmeans_segmentation

IceFloeTracker.landmask

IceFloeTracker.landmask

IceFloeTracker.load_imgs

IceFloeTracker.make_filename

IceFloeTracker.make_hbreak_dict

IceFloeTracker.make_lut

IceFloeTracker.make_psi_s

IceFloeTracker.make_psi_s

IceFloeTracker.makecontourstartatP

IceFloeTracker.mismatch

IceFloeTracker.move

IceFloeTracker.norm

IceFloeTracker.normalize_image

IceFloeTracker.padnhood

IceFloeTracker.prune

IceFloeTracker.regionprops

IceFloeTracker.regionprops_table

IceFloeTracker.remove_padding

IceFloeTracker.segmentation_A

IceFloeTracker.segmentation_B

IceFloeTracker.segmentation_D_E

IceFloeTracker.segmentation_F

IceFloeTracker.segmented_ice_cloudmasking

IceFloeTracker.sharpen

IceFloeTracker.sharpen_gray

IceFloeTracker.timestamp

IceFloeTracker.watershed_ice_floes

IceFloeTracker.@persist